refactored according to snippets formatting. added test file.
